Object pooling system (untested)
Fix client character position not updated
Fix attacks aiming towards 0,0
Fix constant vec provider only accepting ints in object constructor
Remove MousePosVecProvider
UseItemS2C no longer is called when attacking

+++ core/src/main/java/dev/creoii/chaos/util/provider/vecprovider/ConstantVecProvider.java
@@ -26,11 +26,11 @@ public record ConstantVecProvider(NumberProvider x, NumberProvider y) implements
         this(new ConstantNumberProvider(vector2.x), new ConstantNumberProvider(vector2.y));
     }
 
-    public ConstantVecProvider(int x, int y) {
+    public ConstantVecProvider(float x, float y) {
         this(new ConstantNumberProvider(x), new ConstantNumberProvider(y));
     }
 
-    public ConstantVecProvider(int num) {
+    public ConstantVecProvider(float num) {
         this(num, num);
     }
